# iroha-rest-server-swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a REST server to interact with Hyperledger Iroha Blockchain",
        "version": "1.0.0",
        "title": "Iroha REST Server",
        "contact": {
            "email": "dominic.aidoo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http","https"],
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/v1/account/compareAndSetAccountDetail": {
            "post" : {
                "summary" : "Compare and set account detail",
                "description": "Compare and set account detail",
                "produces": ["application/json"],
                "tags":["AccountController"],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Compare and set account detail",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CompareAndSetAccountDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/account/createAccount": {
            "post" : {
                "summary" : "Create account",
                "description": "Create account",
                "produces": ["application/json"],
                "tags":["AccountController"],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Create account",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/account/setAccountDetail": {
            "post" : {
                "summary" : "Set account detail",
                "description": "Set account detail",
                "produces": ["application/json"],
                "tags":["AccountController"],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Set account detail",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SetAccountDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/account/setAccountQuorum": {
            "post" : {
                "summary" : "Set account quorum",
                "description": "Set account quorum",
                "produces": ["application/json"],
                "tags":["AccountController"],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Set account quorum",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SetAccountQuorumRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/account/getAccount/{accountId}": {
            "get" : {
                "summary" : "Get account",
                "description": "Get account",
                "produces": ["application/json"],
                "tags":["AccountController"],
                "parameters": [
                    {
                        "name": "accountId",
                        "in": "path",
                        "type": "string",
                        "description": "Get account",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/account/getAccountAssets": {
            "post" : {
                "summary" : "Get account assets",
                "description": "Get account assets",
                "produces": ["application/json"],
                "tags":["AccountController"],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Get account assets",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/GetAccountAssetsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/account/getRawAccount/{accountId}": {
            "get" : {
                "summary" : "Get raw account",
                "description": "Get raw account",
                "produces": ["application/json"],
                "tags":["AccountController"],
                "parameters": [
                    {
                        "name": "accountId",
                        "in": "path",
                        "type": "string",
                        "description": "Get raw account",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/account/getAccountTransactions": {
            "post" : {
                "summary" : "Get account transactions",
                "description": "Get account transactions",
                "produces": ["application/json"],
                "tags":["AccountController"],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Get account transactions",
                        "required": true,
                        "schema":{
                            "$ref": "#/definitions/GetAccountTransactionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/account/getAccountAssetTransactions": {
            "post" : {
                "summary" : "Get account asset transactions",
                "description": "Get account asset transactions",
                "produces": ["application/json"],
                "tags":["AccountController"],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Get account asset transactions",
                        "required": true,
                        "schema":{
                            "$ref": "#/definitions/GetAccountAssetTransactionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/getAssetInfoResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/IrohaErrorResponse"
                        }
                    }
                }
            }
        },
...
    }, 
    "definitions": {
...
        "CompareAndSetAccountDetailRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                     "type": "string"
                },
                "key": {
                     "type": "string"
                },
                "value": {
                     "type": "string"
                },
                "oldValue": {
                     "type": "string"
                },
                "checkEmpty": {
                    "type": "boolean"
                }
            }
        },
        "CreateAccountRequest": {
            "type": "object",
            "properties": {
                "accountName": {
                     "type": "string"
                },
                "domainId": {
                     "type": "string"
                },
                "publicKey": {
                     "type": "string"
                }
            }
        },
        "SetAccountDetailRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                     "type": "string"
                },
                "key": {
                     "type": "string"
                },
                "value": {
                     "type": "string"
                }
            }
        },
        "SetAccountQuorumRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                     "type": "string"
                },
                "quorum": {
                     "type": "number"
                }
            }
        },
        "GetAccountAssetsRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                     "type": "string"
                },
                "pageSize": {
                     "type": "number"
                },
                "firstAssetId": {
                    "type": "string"
               }
            }
        },
        "GetAccountTransactionsRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                     "type": "string"
                },
                "pageSize": {
                     "type": "number"
                }
            }
        },
        "GetAccountAssetTransactionsRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                     "type": "string"
                },
                "assetId": {
                     "type": "string"
                },
                "pageSize": {
                    "type": "number"
               }
            }
        },
...
        "getAssetInfoResponse": {
            "type": "object",
            "properties": {
                "assetId": {
                     "type": "string"
                },
                "domainId": {
                     "type": "string"
                },
                "precision": {
                    "type": "number"
                }
            }
        },
...
        "IrohaErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                     "type": "string",
                     "example": "Command response error: expected=COMMITTED, actual=REJECTED"
                },
                "status": {
                    "type": "string",
                    "example": "REJECTED"
                }
            }
        },
...
    }
}